unify the tests to create and update daily time records
